# perf: reduce allocation for filename render