--- a/crates/rspack_core/src/options/filename.rs
+++ b/crates/rspack_core/src/options/filename.rs
@@ -10,20 +10,31 @@ use regex::{Captures, Regex};
 use rspack_error::error;
 use rspack_macros::MergeFrom;
 use rspack_util::atom::Atom;
+use rspack_util::ext::CowExt;
 use rspack_util::MergeFrom;
 
 use crate::{parse_resource, AssetInfo, PathData, ResourceParsedData};
 
-pub const FILE_PLACEHOLDER: &str = "[file]";
-pub const BASE_PLACEHOLDER: &str = "[base]";
-pub const NAME_PLACEHOLDER: &str = "[name]";
-pub const PATH_PLACEHOLDER: &str = "[path]";
-pub const EXT_PLACEHOLDER: &str = "[ext]";
-pub const QUERY_PLACEHOLDER: &str = "[query]";
-pub const FRAGMENT_PLACEHOLDER: &str = "[fragment]";
-pub const ID_PLACEHOLDER: &str = "[id]";
-pub const RUNTIME_PLACEHOLDER: &str = "[runtime]";
-pub const URL_PLACEHOLDER: &str = "[url]";
+pub static FILE_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[file]").expect("Should generate regex"));
+pub static BASE_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[base]").expect("Should generate regex"));
+pub static NAME_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[name]").expect("Should generate regex"));
+pub static PATH_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[path]").expect("Should generate regex"));
+pub static EXT_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[ext]").expect("Should generate regex"));
+pub static QUERY_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[query]").expect("Should generate regex"));
+pub static FRAGMENT_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[fragment]").expect("Should generate regex"));
+pub static ID_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[id]").expect("Should generate regex"));
+pub static RUNTIME_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[runtime]").expect("Should generate regex"));
+pub static URL_PLACEHOLDER: Lazy<Regex> =
+  Lazy::new(|| Regex::new("[url]").expect("Should generate regex"));
 pub static HASH_PLACEHOLDER: Lazy<Regex> =
   Lazy::new(|| Regex::new(r"\[hash(:(\d*))?]").expect("Invalid regex"));
 pub static CHUNK_HASH_PLACEHOLDER: Lazy<Regex> =
@@ -179,16 +190,16 @@ impl<F: LocalFilenameFn> Filename<F> {
         Cow::Owned(filename_fn.call(&options, asset_info.as_deref())?)
       }
     };
-    Ok(render_template(template.as_ref(), options, asset_info))
+    Ok(render_template(template, options, asset_info))
   }
 }
 
-fn render_template(
-  template: &str,
+fn render_template<'s>(
+  template: Cow<'s, str>,
   options: PathData,
   mut asset_info: Option<&mut AssetInfo>,
 ) -> String {
-  let mut template = template.to_string();
+  let mut t = template;
   if let Some(filename) = options.filename {
     if let Some(caps) = DATA_URI_REGEX.captures(filename) {
       let ext = mime_guess::get_mime_extensions_str(
@@ -198,48 +209,55 @@ fn render_template(
           .as_str(),
       )
       .map(|exts| exts[0]);
-      template = template.replace(FILE_PLACEHOLDER, "");
-      template = template.replace(QUERY_PLACEHOLDER, "");
-      template = template.replace(FRAGMENT_PLACEHOLDER, "");
-      template = template.replace(PATH_PLACEHOLDER, "");
-      template = template.replace(BASE_PLACEHOLDER, "");
-      template = template.replace(NAME_PLACEHOLDER, "");
-      template = template.replace(
-        EXT_PLACEHOLDER,
-        &ext.map(|ext| format!(".{}", ext)).unwrap_or_default(),
-      );
+      t = t
+        .map(|t| FILE_PLACEHOLDER.replace(t, ""))
+        .map(|t| QUERY_PLACEHOLDER.replace(t, ""))
+        .map(|t| FRAGMENT_PLACEHOLDER.replace(t, ""))
+        .map(|t| PATH_PLACEHOLDER.replace(t, ""))
+        .map(|t| BASE_PLACEHOLDER.replace(t, ""))
+        .map(|t| NAME_PLACEHOLDER.replace(t, ""))
+        .map(|t| {
+          EXT_PLACEHOLDER.replace(t, &ext.map(|ext| format!(".{}", ext)).unwrap_or_default())
+        });
     } else if let Some(ResourceParsedData {
       path: file,
       query,
       fragment,
     }) = parse_resource(filename)
     {
-      template = template.replace(FILE_PLACEHOLDER, &file.to_string_lossy());
-      template = template.replace(
-        EXT_PLACEHOLDER,
-        &file
-          .extension()
-          .map(|p| format!(".{}", p.to_string_lossy()))
-          .unwrap_or_default(),
-      );
+      t = t
+        .map(|t| FILE_PLACEHOLDER.replace(t, file.to_string_lossy()))
+        .map(|t| {
+          EXT_PLACEHOLDER.replace(
+            t,
+            file
+              .extension()
+              .map(|p| format!(".{}", p.to_string_lossy()))
+              .unwrap_or_default(),
+          )
+        });
+
       if let Some(base) = file.file_name().map(|p| p.to_string_lossy()) {
-        template = template.replace(BASE_PLACEHOLDER, &base);
+        t = t.map(|t| BASE_PLACEHOLDER.replace(t, &base));
       }
       if let Some(name) = file.file_stem().map(|p| p.to_string_lossy()) {
-        template = template.replace(NAME_PLACEHOLDER, &name);
+        t = t.map(|t| NAME_PLACEHOLDER.replace(t, &name));
       }
-      template = template.replace(
-        PATH_PLACEHOLDER,
-        &file
-          .parent()
-          .map(|p| p.to_string_lossy())
-          // "" -> "", "folder" -> "folder/"
-          .filter(|p| !p.is_empty())
-          .map(|p| p + "/")
-          .unwrap_or_default(),
-      );
-      template = template.replace(QUERY_PLACEHOLDER, &query.unwrap_or_default());
-      template = template.replace(FRAGMENT_PLACEHOLDER, &fragment.unwrap_or_default());
+      t = t
+        .map(|t| {
+          PATH_PLACEHOLDER.replace(
+            t,
+            &file
+              .parent()
+              .map(|p| p.to_string_lossy())
+              // "" -> "", "folder" -> "folder/"
+              .filter(|p| !p.is_empty())
+              .map(|p| p + "/")
+              .unwrap_or_default(),
+          )
+        })
+        .map(|t| QUERY_PLACEHOLDER.replace(t, &query.unwrap_or_default()))
+        .map(|t| FRAGMENT_PLACEHOLDER.replace(&t, &fragment.unwrap_or_default()));
     }
   }
   if let Some(content_hash) = options.content_hash {
@@ -247,8 +265,8 @@ fn render_template(
       // set version as content hash
       asset_info.version = content_hash.to_string();
     }
-    template = CONTENT_HASH_PLACEHOLDER
-      .replace_all(&template, |caps: &Captures| {
+    t = t.map(|t| {
+      CONTENT_HASH_PLACEHOLDER.replace_all(t, |caps: &Captures| {
         let content_hash = &content_hash[..hash_len(content_hash, caps)];
         if let Some(asset_info) = asset_info.as_mut() {
           asset_info.set_immutable(true);
@@ -256,12 +274,12 @@ fn render_template(
         }
         content_hash
       })
-      .into_owned();
+    });
   }
   if let Some(hash) = options.hash {
     for reg in [&HASH_PLACEHOLDER, &FULL_HASH_PLACEHOLDER] {
-      template = reg
-        .replace_all(&template, |caps: &Captures| {
+      t = t.map(|t| {
+        reg.replace_all(t, |caps: &Captures| {
           let hash = &hash[..hash_len(hash, caps)];
           if let Some(asset_info) = asset_info.as_mut() {
             asset_info.set_immutable(true);
@@ -269,21 +287,21 @@ fn render_template(
           }
           hash
         })
-        .into_owned();
+      });
     }
   }
   if let Some(chunk) = options.chunk {
     if let Some(id) = &options.id {
-      template = template.replace(ID_PLACEHOLDER, id);
+      t = t.map(|t| ID_PLACEHOLDER.replace(t, *id));
     } else if let Some(id) = &chunk.id {
-      template = template.replace(ID_PLACEHOLDER, &id.to_string());
+      t = t.map(|t| ID_PLACEHOLDER.replace(t, id));
     }
     if let Some(name) = chunk.name_for_filename_template() {
-      template = template.replace(NAME_PLACEHOLDER, name);
+      t = t.map(|t| NAME_PLACEHOLDER.replace(t, name));
     }
     if let Some(d) = chunk.rendered_hash.as_ref() {
-      template = CHUNK_HASH_PLACEHOLDER
-        .replace_all(&template, |caps: &Captures| {
+      t = t.map(|t| {
+        CHUNK_HASH_PLACEHOLDER.replace_all(t, |caps: &Captures| {
           let hash = &**d;
           let hash = &hash[..hash_len(hash, caps)];
           if let Some(asset_info) = asset_info.as_mut() {
@@ -292,22 +310,22 @@ fn render_template(
           }
           hash
         })
-        .into_owned();
+      });
     }
   }
 
   if let Some(id) = &options.id {
-    template = template.replace(ID_PLACEHOLDER, id);
+    t = t.map(|t| ID_PLACEHOLDER.replace(t, *id));
   } else if let Some(module) = options.module {
     if let Some(chunk_graph) = options.chunk_graph {
       if let Some(id) = chunk_graph.get_module_id(module.identifier()) {
-        template = template.replace(ID_PLACEHOLDER, id);
+        t = t.map(|t| ID_PLACEHOLDER.replace(t, id));
       }
     }
   }
-  template = template.replace(RUNTIME_PLACEHOLDER, options.runtime.unwrap_or("_"));
+  t = t.map(|t| RUNTIME_PLACEHOLDER.replace(t, options.runtime.unwrap_or("_")));
   if let Some(url) = options.url {
-    template = template.replace(URL_PLACEHOLDER, url);
+    t = t.map(|t| URL_PLACEHOLDER.replace(t, url));
   }
-  template
+  t.into_owned()
 }