<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

`String::replace` allocates `String` on each iteration and `Regex::replace_all` only returns `Cow::Owned(...)` if there's an actual replacement takes place. Thus, we are moving from `String::replace` to `Regex::replace_all`.

For test cases: 
- 0: "[id].[name].css"
- 1: "[id].[name].[fullhash][ext][[query]"
- 2: "[id].[name].[fullhash][ext][[query][id].[name].[fullhash][ext][[query]"

**Case: "[id].[name].css"** (22% faster)
```
Replace/String::replace/0
                        time:   [31.403 µs 31.468 µs 31.529 µs]
                        change: [+0.3228% +0.7846% +1.2319%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
Replace/Regex::replace/0
                        time:   [24.505 µs 24.594 µs 24.692 µs]
                        change: [-0.0425% +0.2968% +0.6560%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
```

**Case: "[id].[name].[fullhash][ext][[query]"** (32% faster)
```
Replace/String::replace/1
                        time:   [40.020 µs 40.107 µs 40.190 µs]
Replace/Regex::replace/1
                        time:   [26.951 µs 27.099 µs 27.277 µs]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
```

**Case: "[id].[name].[fullhash][ext][[query][id].[name].[fullhash][ext][[query]"** (42% faster)
```
Replace/String::replace/2
                        time:   [54.491 µs 54.649 µs 54.795 µs]
Replace/Regex::replace/2
                        time:   [31.855 µs 31.955 µs 32.060 µs]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

![Replace benchmark lines](https://github.com/user-attachments/assets/d72f8a1e-00ec-4f4d-9d58-acc124728546)



<!-- Can you explain the reasoning behind implementing this change? What problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).